# social-network-hw3/openapi.yaml
openapi: 3.0.3
info:
  title: Social Network Users API
  description: API для регистрации, аутентификации пользователей и управления постами
  version: "1.0"
paths:
  /register:
    post:
      summary: Регистрация нового пользователя
      requestBody:
        description: Данные для регистрации
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Ошибка регистрации
  v1/login:
    post:
      summary: Аутентификация пользователя
      requestBody:
        description: Логин и пароль пользователя
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Пользователь успешно аутентифицирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
        '400':
          description: Неверные учетные данные
  /profile:
    get:
      summary: Получение данных профиля
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Неавторизованный
    put:
      summary: Обновление данных профиля
      security:
        - bearerAuth: []
      requestBody:
        description: Данные для обновления профиля
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Профиль успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Неавторизованный
  /posts:
    post:
      summary: Создание нового поста
      security:
        - bearerAuth: []
      requestBody:
        description: Данные нового поста
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: Пост успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Ошибка валидации
    get:
      summary: Получение списка постов
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список постов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

  /posts/{post_id}:
    get:
      summary: Получение поста по ID
      security:
        - bearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Пост не найден

    put:
      summary: Обновление поста
      security:
        - bearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Обновлённые данные поста
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Пост обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Ошибка валидации
        '404':
          description: Пост не найден

    delete:
      summary: Удаление поста
      security:
        - bearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Пост успешно удалён
        '404':
          description: Пост не найден

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    UserRegister:
      type: object
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
        email:
          type: string
          format: email
      required:
        - login
        - password
        - email
    UserLogin:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        phone_number:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        email:
          type: string
          format: email
        phone_number:
          type: string
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        is_private:
          type: boolean
        author_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PostCreate:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        is_private:
          type: boolean
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        is_private:
          type: boolean